Modernize Utils to Java 21, make them so they can't be instantiated.

+++ TNPC-API/src/main/java/net/tnemc/plugincore/api/server/player/PlayerProvider.java
@@ -46,7 +46,7 @@ public interface PlayerProvider {
 
     int getExp();
 
-    void setExp(int exp);
+    void setExp(float exp);
 
     int getExpLevel();
